% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gcd.R
\name{calculate_gcd}
\alias{calculate_gcd}
\title{Calculate Great Circle Distance}
\usage{
calculate_gcd(long_1, lat_1, long_2, lat_2)
}
\arguments{
\item{long_1}{Longitude of first coordinate, in radians.}

\item{lat_1}{Latitude of first coordinate, in radians.}

\item{long_2}{Longitude of second coordinate, in radians.}

\item{lat_2}{Latitude of second coordinate, in radians.}
}
\value{
A numeric vector of distance between the two given coordinates in km.
}
\description{
Calculate Great Circle Distance
}
\examples{
calculate_gcd <- function(deg2rad(-6), deg2rad(55),
                          deg2rad(-7), deg2rad(56))
}
