% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg2rad.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert Degrees To Radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{numeric degrees (0-360)}
}
\value{
numeric radians
}
\description{
Convert Degrees To Radians
}
\examples{
deg2rad(90)
deg2rad(45)

}
