% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_campus_location.R
\name{get_campus_location}
\alias{get_campus_location}
\title{Get Campus Location}
\usage{
get_campus_location(campus_name, campus_file_path)
}
\arguments{
\item{campus_name}{Name of campus that is to be returned.}

\item{campus_file_path}{Path to file containing all campuses and locations, must be .csv.}
}
\value{
A data frame with one row and values \code{latitude} and \code{longitude}.
}
\description{
Get Campus Location
}
\examples{
get_campus_location("Indeed Dublin HQ", "./path_to_campus_file.csv")
}
