% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_customer_list.R
\name{import_customer_list}
\alias{import_customer_list}
\title{Import Customer Data From File To Data Frame}
\usage{
import_customer_list(customers_path)
}
\arguments{
\item{customers_path}{String containing path to json file}
}
\value{
A data frame with
}
\description{
Import Customer Data From File To Data Frame
}
\examples{
\dontrun{import_customer_list("./customer_data/customers.txt")}
}
